+++ packages/analytics-js-common/src/utilities/timestamp.ts
@@ -1,10 +1,9 @@
+const getFormattedTimestamp = (date: Date): string => date.toISOString();
+
 /**
  * To get the current timestamp in ISO string format
  * @returns ISO formatted timestamp string
  */
-const getCurrentTimeFormatted = (): string => {
-  const curDateTime = new Date().toISOString();
-  return curDateTime;
-};
+const getCurrentTimeFormatted = (): string => getFormattedTimestamp(new Date());
 
-export { getCurrentTimeFormatted };
+export { getCurrentTimeFormatted, getFormattedTimestamp };