+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.48.23](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.48.22...rudder-sdk-js@2.48.23) (2024-10-25)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.11.0`
 ## [2.48.22](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.48.21...rudder-sdk-js@2.48.22) (2024-10-21)
 
 ### Dependency Updates